Fixes to old code after updating to JDK 11.
One test of laziness in PackageGrouping had a comment "don't understand why this is 2 not 1", and now it is 1. Changed enforcement.

One test of Circuit now revealed a ConcurrentModificationException when doing computeIfAbsent on a HashMap. It turns out in JDK 9+ the computation must not modify the map. Changed this to use a Guava LoadingCache instead.